# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Qwerty Learner API
  description: |
    前后端分离二次开发接口规范。
    对齐前端类型：`src/typings/*`、`src/utils/db/record.ts`。
    章节长度 CHAPTER_LENGTH = 20；时间戳为 Unix 秒（UTC）。
  version: 1.0.0
  contact:
    name: Qwerty Learner
servers:
  - url: /api/v1
    description: Relative to API host
...
paths:
...
  /dictionaries:
    get:
      tags: [Dictionaries]
      summary: 词典列表
      operationId: listDictionaries
      parameters:
        - $ref: '#/components/parameters/languageCategory'
        - name: category
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: q
          in: query
          schema:
            type: string
          description: 名称/描述模糊搜索
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DictionaryListResponse'
    post:
      tags: [Dictionaries]
      summary: 创建词典（管理员）
      operationId: createDictionary
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DictionaryCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dictionary'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /reviews:
    get:
      tags: [Reviews]
      summary: 复习记录列表
      operationId: listReviews
      security:
        - bearerAuth: []
      parameters:
        - name: dict
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/pageSize'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ReviewRecord'
                  total:
                    type: integer
    post:
      tags: [Reviews]
      summary: 基于错题生成新复习
      description: |
        排序：errorCountScore*0.6 + latestErrorTimeScore*0.4，升序。
        与 `src/utils/db/review-record.ts` 一致。
      operationId: createReview
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewRecord'
...
  /stats/chapters:
    get:
      tags: [Stats]
      summary: 章节统计
      operationId: getChapterStats
      security:
        - bearerAuth: []
      parameters:
        - name: dict
          in: query
          required: true
          schema:
            type: string
        - name: chapter
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChapterStats'
  /stats/dictionaries/{dictId}:
    get:
      tags: [Stats]
      summary: 词典练习进度
      operationId: getDictStats
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/dictId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DictStats'
  /stats/analysis:
    get:
      tags: [Stats]
      summary: 分析页总览
      operationId: getAnalysisStats
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/from'
        - $ref: '#/components/parameters/to'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisStats'
  /stats/summary:
    get:
      tags: [Stats]
      summary: 用户总览
      operationId: getStatsSummary
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsSummary'
  /settings:
    get:
      tags: [Settings]
      summary: 获取用户设置
      operationId: getSettings
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSettings'
    put:
      tags: [Settings]
      summary: 全量更新设置
      operationId: putSettings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSettings'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSettings'
    patch:
      tags: [Settings]
      summary: 部分更新设置
      operationId: patchSettings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSettingsPatch'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSettings'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    dictId:
      name: dictId
      in: path
      required: true
      schema:
        type: string
        example: cet4
    languageCategory:
      name: languageCategory
      in: query
      schema:
        $ref: '#/components/schemas/LanguageCategoryType'
    from:
      name: from
      in: query
      schema:
        type: integer
      description: Unix 秒起始
    to:
      name: to
      in: query
      schema:
        type: integer
      description: Unix 秒结束
    page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    pageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    PronunciationType:
      type: string
      enum: [us, uk, romaji, zh, ja, de, hapin, kk, id]
    PhoneticType:
      type: string
      enum: [us, uk, romaji, zh, ja, de, hapin, kk, id]
    LanguageType:
      type: string
      enum: [en, romaji, zh, ja, code, de, kk, hapin, id]
    LanguageCategoryType:
      type: string
      enum: [en, ja, de, code, kk, id]
    WordDictationType:
      type: string
      enum: [hideAll, hideVowel, hideConsonant, randomHide]
    WordDictationOpenBy:
      type: string
      enum: [user, auto]
...
    Word:
      type: object
      required: [name, trans]
      properties:
        name:
          type: string
        trans:
          type: array
          items:
            type: string
        usphone:
          type: string
          default: ''
        ukphone:
          type: string
          default: ''
        notation:
          type: string
          nullable: true
...
    Dictionary:
      type: object
      required: [id, name, description, category, tags, length, chapterCount, language, languageCategory]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        length:
          type: integer
        chapterCount:
          type: integer
          description: ceil(length / 20)
        language:
          $ref: '#/components/schemas/LanguageType'
        languageCategory:
          $ref: '#/components/schemas/LanguageCategoryType'
        defaultPronIndex:
          type: integer
          nullable: true
    DictionaryCreate:
      type: object
      required: [id, name, description, category, tags, language, languageCategory]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        language:
          $ref: '#/components/schemas/LanguageType'
        languageCategory:
          $ref: '#/components/schemas/LanguageCategoryType'
        defaultPronIndex:
          type: integer
          nullable: true
    DictionaryListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Dictionary'
        total:
          type: integer
...
    ReviewRecord:
      type: object
      properties:
        id:
          type: integer
        dict:
          type: string
        index:
          type: integer
        createTime:
          type: integer
        isFinished:
          type: boolean
        words:
          type: array
          items:
            $ref: '#/components/schemas/Word'
    ReviewCreate:
      type: object
      required: [dict]
      properties:
        dict:
          type: string
        errorData:
          type: array
          description: 可选；不传则服务端自行聚合错题
          items:
            type: object
            required: [word, errorCount, latestErrorTime, originData]
            properties:
              word:
                type: string
              errorCount:
                type: integer
              latestErrorTime:
                type: integer
              originData:
                $ref: '#/components/schemas/Word'
...
    ChapterStats:
      type: object
      properties:
        dict:
          type: string
        chapter:
          type: integer
        exerciseCount:
          type: integer
        avgWrongWordCount:
          type: number
        avgWrongInputCount:
          type: number
    DictStats:
      type: object
      properties:
        dict:
          type: string
        exercisedChapterCount:
          type: integer
        chapterCount:
          type: integer
    ActivityDay:
      type: object
      properties:
        date:
          type: string
          format: date
        count:
          type: integer
        level:
          type: integer
          minimum: 0
          maximum: 4
    AnalysisStats:
      type: object
      properties:
        isEmpty:
          type: boolean
        exerciseRecord:
          type: array
          items:
            $ref: '#/components/schemas/ActivityDay'
        wordRecord:
          type: array
          items:
            $ref: '#/components/schemas/ActivityDay'
        wpmRecord:
          type: array
          items:
            type: array
            minItems: 2
            maxItems: 2
            items: {}
          description: '[date, wpm][]'
        accuracyRecord:
          type: array
          items:
            type: array
            minItems: 2
            maxItems: 2
            items: {}
        wrongTimeRecord:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              value:
                type: integer
    StatsSummary:
      type: object
      properties:
        wordRecordCount:
          type: integer
        chapterRecordCount:
          type: integer
        totalTimeSeconds:
          type: integer
        firstPracticeAt:
          type: integer
          nullable: true
    SoundResource:
      type: object
      properties:
        key:
          type: string
        name:
          type: string
        filename:
          type: string
    UserSettings:
      type: object
      properties:
        currentDict:
          type: string
        currentChapter:
          type: integer
        loopWordConfig:
          type: object
          properties:
            times:
              type: number
        keySoundsConfig:
          type: object
          properties:
            isOpen:
              type: boolean
            isOpenClickSound:
              type: boolean
            volume:
              type: number
            resource:
              nullable: true
              allOf:
                - $ref: '#/components/schemas/SoundResource'
        hintSoundsConfig:
          type: object
          properties:
            isOpen:
              type: boolean
            volume:
              type: number
            isOpenWrongSound:
              type: boolean
            isOpenCorrectSound:
              type: boolean
            wrongResource:
              nullable: true
              allOf:
                - $ref: '#/components/schemas/SoundResource'
            correctResource:
              nullable: true
              allOf:
                - $ref: '#/components/schemas/SoundResource'
        pronunciation:
          type: object
          properties:
            isOpen:
              type: boolean
            volume:
              type: number
            type:
              $ref: '#/components/schemas/PronunciationType'
            name:
              type: string
            isLoop:
              type: boolean
            isTransRead:
              type: boolean
            transVolume:
              type: number
            rate:
              type: number
        fontsize:
          type: object
          properties:
            foreignFont:
              type: number
            translateFont:
              type: number
        randomConfig:
          type: object
          properties:
            isOpen:
              type: boolean
        phoneticConfig:
          type: object
          properties:
            isOpen:
              type: boolean
            type:
              $ref: '#/components/schemas/PhoneticType'
        wordDictationConfig:
          type: object
          properties:
            isOpen:
              type: boolean
            type:
              $ref: '#/components/schemas/WordDictationType'
            openBy:
              $ref: '#/components/schemas/WordDictationOpenBy'
        isShowPrevAndNextWord:
          type: boolean
        isIgnoreCase:
          type: boolean
        isShowAnswerOnHover:
          type: boolean
        isTextSelectable:
          type: boolean
        isOpenDarkMode:
          type: boolean
        dismissStartCardDate:
          type: string
          format: date-time
          nullable: true
        hasSeenEnhancedPromotion:
          type: boolean
    UserSettingsPatch:
      type: object
      additionalProperties: true
      description: 任意 UserSettings 子集，对象字段深合并